backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
celery
redis
python-docx
python-pptx
langgraph
langchain-google-genai
python-multipart
pyjwt
python-jose[cryptography]
passlib
bcrypt
httpx
gTTS
reportlab
